# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Frames API
  version: 0.1.0
servers:
  - url: http://localhost:3001
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /admin/overview:
    get:
      responses:
        "200": { description: Administrator aggregate overview }
        "403": { description: Administrator access required }
...
  /admin/users/{id}:
    get:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Administrator user detail }
  /admin/users/{id}/role:
    patch:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Administrator role update }
...
  /auth/register:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, displayName, email, password]
              properties:
                username: { type: string }
                displayName: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        "201":
          description: Created user and tokens
  /auth/login:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Authenticated user and tokens
...
  /users/{username}:
    get:
      parameters:
        - in: path
          name: username
          required: true
          schema: { type: string }
      responses:
        "200": { description: User profile }
...
  /friendships/{id}:
    patch:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Friendship accepted or blocked }
    delete:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204": { description: Friendship removed }
...
  /archive/daily/{date}:
    get:
      parameters:
        - in: path
          name: date
          required: true
          schema: { type: string, format: date }
      responses:
        "200": { description: Daily frame }
...
  /search/archive:
    get:
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
      responses:
        "200": { description: Archive search results }
...